requirements.txt
pdfminer.six==20221105
python-docx==0.8.11
nltk==3.8.1
pandas==1.4.2
numpy==1.22.3
scikit-learn==1.0.2
seaborn==0.11.2
matplotlib==3.5.1 